+++ src/index.js
@@ -1,5 +1,5 @@
 import ReactDOM from "react-dom/client";
-import App from "./component/App";
+import App from "./App";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(<App />);
